requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
psycopg2-binary==3.2.3
starlette==0.41.3
typing-extensions==4.12.2
dotenv==0.0.5